requirements.txt
python-dotenv
json
requests
os